Only use ij ion symmetry for H_ab, don't use ab symmetry for kvecs
Using ij ion symmetry for H_ab allows the stored value of H_ab to be
much smaller, without affecting performance, so keep this optimisation.
However, it's been found that using the fact that kvecs_ab is symmetric
in alpha, beta complicates the code without resulting in much (if any)
performance improvement.