Sphere sampling of neutron scattering function
Another function is provided, similar to sample_sphere_dos, that
samples the structure factor and returns a StructureFactor object.

As a convenience, it generates Debye-Waller data on-the-fly if this is
not provided.

The StructureFactor is not such a convenient object for broadening and
plotting as a Spectrum1D. Perhaps would should do the binning within
this function - but what if the user does want the intermediate
object?