Reference data: allow files to be specified as collection
The "collection" argument to get_reference_data will now accept the
path to a JSON file of the same format as those in euphonic.data.

In the unlikely event that the user passes a filename that is also a
key for a standard data set from euphonic.data, the inbuilt data will
take precedence. Users are recommended to use a .json extension in any
case, but would also be able to work around the clash by specifying
an absolute path.

The new mechanism is tested using some fictitious data in a temporary file.