Reference data system (#79), closes #78
* Reference data system: json files, importlib, units & complex nums

An initial implementation of a reference data system:

- Reference data is sorted by label and property
  - This allows trusted collections of multiple properties to be
    accessed systematically
  - Currently the labels and properties are related to files in a
    hard-coded dict stored inside the lookup function
  - It would be good to move this lookup table to somewhere more
    convenient for editing/updating/reference
  - It would also be nice to provide a facility for direct access to
    files outside the Euphonic package

- A minor JSON extension is implemented to include complex numbers and
  unit metadata.

* Add unit test for reference data loading

* Reference data: better names "collection" and "physical_property"

Thanks @rebeccafair and @mducle

* Flake8 cleanup of euphonic.util

* Reference data: rework so multiple properties can be stored per file

This will allow loading from data folder or a file path with
consistent notation.

* Reference data: allow files to be specified as collection

The "collection" argument to get_reference_data will now accept the
path to a JSON file of the same format as those in euphonic.data.

In the unlikely event that the user passes a filename that is also a
key for a standard data set from euphonic.data, the inbuilt data will
take precedence. Users are recommended to use a .json extension in any
case, but would also be able to work around the clash by specifying
an absolute path.

The new mechanism is tested using some fictitious data in a temporary file.

* Ref data: type hints, docstrings, add to calculate_structure_factor

* Structure factor from ref data: added tests

* Doc updates for reference data system; add sphinx-autodoc-typehints

* CHANGELOG update

* Ref data docs cleanup

* Pin sphinx package versions

By default RTD uses an old sphinx that is not compatible with
autodoc-typehints.