Implement averaged structure factor to Spectrum1D
A similar method to calculate_sqw_map() is provided to
StructureFactor, which returns a 1D average over q-points. This is
appropriate for spherical averaging or Brillouin-zone sums. Most of
the code is shared in a private method.

A "temperature" argument is exposed to sample_sphere_structure_factor
so that appropriate corrections may be requested.