Spherical averaging tools (#73)
* Begin implementing spherical sampling functions

- Sampling functions go in main library under euphonic.sampling
- A demonstration script is added to the Euphonic scripts folder.
  - This script enables one to conveniently visualise different
  sampling schemes for debugging and discussion purposes.
  - For now it is not added to the setup.py installation, nor is it
    clear that it ever should be.

* Spherical sampling: improved polar-coordinates sampling

The maths for this should be derived more neatly and layed out at some
point, but it seems to work!

In polar coordinates, a series of constant-theta rings are filled with
evenly-spaced points by subdividing phi. However, each ring has its
own number of points, in order to give a fairly uniform distribution.

The heuristic is obtained by making the distance of "longitude step"
between rings equal to the average spacing of points if all rings were
unfolded, set end-to-end and uniformly filled with the requested NPTS.
The average circumference along a latitude line is 4 (by integrating
sin(x)). To satisfy the distance constraint with radius r and n_theta
divisions:

(4 r n_theta / npts) = (pi r / n_theta)

which leads to

n_theta = sqrt( pi / 4 npts )

Which is going to be irrational for any integer npts, so a bit of
rounding is required. The results actually look pretty nice, though!

The method could probably be improved by some thoughtful orientation
in phi; currently all "latitudes" include a point at phi=0, creating a
consistent line of longitude. Such alignment could be broken by a
Golden ratio rotation at each theta.

* Sphere sampling: refactor, add random

* Spherical sampling: add square grid projection, jitter options

* Spherical sampling: function from ForceConstants -> DOS

This code was adapted from scripts used for qualitative demonstrations
of sampling methods. DOS sampling was implemented initially in order
to examine convergence without Debye-Waller effects.

- At this point an explicit npts is provided. This is likely to be
  developed/replaced/wrapped with a more adaptive scheme.

- This function only samples at the surface of a sphere; schemes that
  include a range of mod(q) within a bin will be implemented as separate
  function(s) as they are defined by a substantially different/longer
  set of parameters.

* Sphere sampling of neutron scattering function

Another function is provided, similar to sample_sphere_dos, that
samples the structure factor and returns a StructureFactor object.

As a convenience, it generates Debye-Waller data on-the-fly if this is
not provided.

The StructureFactor is not such a convenient object for broadening and
plotting as a Spectrum1D. Perhaps would should do the binning within
this function - but what if the user does want the intermediate
object?

* Update reference data syntax

* Implement averaged structure factor to Spectrum1D

A similar method to calculate_sqw_map() is provided to
StructureFactor, which returns a 1D average over q-points. This is
appropriate for spherical averaging or Brillouin-zone sums. Most of
the code is shared in a private method.

A "temperature" argument is exposed to sample_sphere_structure_factor
so that appropriate corrections may be requested.

* Sampling: first test (golden square)

* Spherical sampling: more tests

* Complete test coverage of euphonic.sampling; SPI min points error

- Use a common function to tidy tests and test all sphere sampling
  methods against previously calculated data. This has not been
  verified against an external reference, but should at least catch
  unexpected changes.

- The "spherical polar improved" scheme hits indexing problems if less
  than six points are used. Test this and give a more meaningful
  message.

- Because of the way Python handles exceptions in generators, this
  error is not actually raised until the generator is called. If
  necessary this can be worked around by wrapping the generator code
  in another function and calling it after the npts check. In this
  case it doesn't look like a big problem so leave the cleaner code in place.

* Powder sampling tests: mock out sampling functions

* Powder: multiply sphere by mod(q); add tests for sphere DOS sampling

- mod(q) dependence of sampling sphere was not yet implemented
- this needs further work to convert between fractional/Cartesian
  coordinates to sample a proper "sphere"
- Meanwhile, use mocking to check sphere DOS function is making
  sampling calls as expected

* Powder averaging: implement and test frac/Cart coords conversion

Add a private method for solving fractional coordinates from Cartesian
input (e.g. a constant-mod(q) sphere). Test that this behaves as expected.

* Pytest refactoring for powder DOS test; clean up units

* Sphere sampling structure factor: mocking test setup

A few of the internals still need testing

* euphonic.powder testing: PEP8 cleanup

* Powder scattering function tests: more mock sanity-checking

* Powder testing: complete coverage of sample_sphere_structure_factor

- This involves some complicated mocking of pretty much every class in
  Euphonic
- but means we avoid relying on numerical tests
- and helps narrow down problems when they sit in those other classes/functions

Just a few more bits of this module to test!

* Test remaining parts of euphonic.powder

* Docs update: euphonic.powder

- Add to API section on tutorials page
- Change Google-style docstrings to numpy-style and fix formatting issues

* Docs update: euphonic.sampling

- Use numpy-style docstrings
- Create RST sphinx page and add to API index in tutorials.rst

* Powder averaging: tweaks in response to review comments

* Rename sphere sampling visualistion script, add to setup.py

The name does not follow the existing convention; it was agreed that
the others would be named in the near future.

* Sphere sampling script tests

- Script test is added for euphonic_sphere_sampling, imitating design
  of other script tests
- Script name changed to match python conventions in order to enable
  import. Cleaning this up is a motivation to sort out the script
  entry points in general!
- Script slightly refactored to accept parser arguments via Python
- ax.scatter used for 2D plots to match 3D and simplify test logic

* Changelog update: spherical averaging features

* Changelog fix: markdown bullets don't always do what you expect