Spectrum classes: Spectrum1DCollection, Spectrum base class
This setup for a spectrum base class is a bit of a hack: rename
Spectrum1D to Spectrum, and inherit it to Spectrum1D. The purpose is
to make sure that isinstance(spectrum, Spectrum1D) will FAIL if the
spectrum is a Spectrum2D.

Ideally we would have some shared code in a Spectrum ABC, and unique
features in each subclass. But because there is a lot of "magic"
around setting the properties of these classes with appropriate units,
it is not so straightforward to establish abstract properties.