Spectrum1D split(): Expand test coverage
This now gives full coverage for the base Spectrum class and Spectrum1D
class.

The new tests are a bit verbose compared to the existing setup,
because the sample data is generated on the fly rather than read from
JSON. At this point while features and tests are being developed this
is a better workflow; it can be tidied up into JSON regression tests
once we are satisfied with the functionality.