Add function to get key for script test outputs
Now that the input file is included in dos_args, dispersion_args etc.,
it cannot be used naively to produce a key with str.join, as this
includes the absolute path, which will be different on different
systems. Instead extract the directory name and filename and use those
in the key.