Powder map tool: add performance settings, use C ext by default
There's no compelling reason to have use_c=False outside of testing as
it will fall through if the extension is unavailable. This is being
looked at; in the meantime, let's have it on by default for CLI users,
with the option to disable.

Threading currently has to be set explicitly. As we are likely to
improve the defaults here, I have the CLI logic leaving this option
unset if it is not used. That way defaults sitting deeper in Euphonic
will still work, but users have the option to override if they are
doing something clever.