Refactor command-line parsers
Group common features by keywords which can be passed to
_get_cli_parser for consistent sorting and formatting.

Parser is returned with argument groups (which do not have a 1:1 map
with feature keywords), so that unique options can be added to a
relevant section of the help page.

This is quite a drastic refactor! Creating consistent feature groups
also highlighted omitted/inconsistent features.

- Debye-Waller corrections can now be added when using intensity-map.
- q-distance is renamed to q-spacing
- euphonic-dos now has a default sampling density rather than a default
  M-P grid. (This is a better default as it will adapt to the unit
  cell size.)

One test input was tweaked to demonstrate that the results are
consistent, and should be updated to better represent the new feature
set:

- euphonic-dos "all defaults from force constants" test was changed to
  specify a 6x6x6 grid (the former default)