Reduce number of loops calculating dyn mat in C
Only the upper triangular needs to be calculated, as both
the dynamical matrix and its gradients are Hermitian. When initially
implementing the gradients calculation, the whole matrix was calculated
for simplicity. Now that it has been tested, only the upper triangular
needs to be calculated, and the rest filled in. This brings the
dynamical matrix calculation performance back into line with what is
implemented in master.