Introduce auto MP grids; fix powder tests for QpointFrequencies
The DOS tests were breaking as they expected
calculate_qpoint_phonon_modes to be called, and this has changed to
use calculate_frequencies instead.

So far the logic for dispatching between alternative grid
specifications (grid vs density) remains in the command-line tools,
but the ability to select a mesh based on density is now attached as a
method on Crystal, as this is where the reciprocal lattice information
is guaranteed to exist. This is easily passed to the mp_grid() utility
function.