Add command-line tool for powder-averaged 2D spectra (#127)
* Powder map CLI tool: fixed-npts, DOS and coherent S supported

* Powder map tool: implement npts/q scaling, tqdm progress bar

* Powder map tool: add performance settings, use C ext by default

There's no compelling reason to have use_c=False outside of testing as
it will fall through if the extension is unavailable. This is being
looked at; in the meantime, let's have it on by default for CLI users,
with the option to disable.

Threading currently has to be set explicitly. As we are likely to
improve the defaults here, I have the CLI logic leaving this option
unset if it is not used. That way defaults sitting deeper in Euphonic
will still work, but users have the option to override if they are
doing something clever.

* Add powder map CLI to setup.py

The name "euphonic-powder-map" is still open to debate!

* Refactor command-line parsers

Group common features by keywords which can be passed to
_get_cli_parser for consistent sorting and formatting.

Parser is returned with argument groups (which do not have a 1:1 map
with feature keywords), so that unique options can be added to a
relevant section of the help page.

This is quite a drastic refactor! Creating consistent feature groups
also highlighted omitted/inconsistent features.

- Debye-Waller corrections can now be added when using intensity-map.
- q-distance is renamed to q-spacing
- euphonic-dos now has a default sampling density rather than a default
  M-P grid. (This is a better default as it will adapt to the unit
  cell size.)

One test input was tweaked to demonstrate that the results are
consistent, and should be updated to better represent the new feature
set:

- euphonic-dos "all defaults from force constants" test was changed to
  specify a 6x6x6 grid (the former default)

* Further CLI tweaks: expose eta-scale, tidy up kwargs to calc modes

* Tweak DW/temperature options in powder module to match CLI

This was broken as function couldn't deal with a "None" temperature,
which is now the preferred way of disabling D-W term.

* Add script tests for euphonic-powder-map

* Tweak force constants qpt-reduction fallback for old Numpy

This was falling over in the new powder script tests, because np.view
doesn't accept non-contiguous arrays. The associated error message
does not make it very clear that this is the problem: "new type not
compatible with array".

* Powder map script docs

* Show defaults in CLI -h pages

There's an argparse feature for that!

* Powder CLI review comments: quick fixes

* Add a powder-map test for --temperature argument

* Use QpointFrequencies where possible in CLI tools

* Update intensity-map test to check --temperature option

This branch managed to break the option without breaking the tests, so
better make sure that's covered.

* pytest isn't a dependency; check "== approx()" with np.isclose

This was breaking the docs build

* Introduce auto MP grids; fix powder tests for QpointFrequencies

The DOS tests were breaking as they expected
calculate_qpoint_phonon_modes to be called, and this has changed to
use calculate_frequencies instead.

So far the logic for dispatching between alternative grid
specifications (grid vs density) remains in the command-line tools,
but the ability to select a mesh based on density is now attached as a
method on Crystal, as this is where the reciprocal lattice information
is guaranteed to exist. This is easily passed to the mp_grid() utility
function.

* Expose MP sampling density as option for powder S(|q|,w)

Hard-coded numbers are bad!

* Update euphonic.powder tests to check auto MP grid args are handled

* Add powder-map test for --temperature=1000 --weights=coherent

* Rename auto-MP-grid functions, add test for Crystal method

* Changelog update for powder CLI and related refactor

* euphonic-dos: replace old defaults test with a --grid-spacing test

After verifying that the replaced test was passing with --grid set
equivalent to the former defaults.