Fix PDOS normalisation
Using density=True in np.histogram normalises the area to 1,
which is incorrect when computing PDOS when the total DOS should
be 1 and each PDOS should be a fraction of that. Instead use digitize
to enable weighting of each individual frequency to be binned, taking
into account bin width. Rely on the input mode_weights to give the
correct overall area (not necessarily 1). This method will also
weight correctly if some frequencies are outside the bin range.