Fix bug in powder-map energy units
The previous implementation for getting energy bins for CLI tools
implicitly relied on modes.frequencies and energy_unit being in
the same units, so would silently give incorrect results if
args.emin and args.emax were not defined, and
modes.frequencies_unit = energy_unit was not called first. This was
not called in powder-map so ended up with energy bins being created
with incorrect limits. Have refactored this so that energy bins will
be generated in the same units as modes to reduce some complexity,
and test data has been regenerated for the offending test cases.
Some verification logic for the energy units has also been removed,
but this should already be handled by the verification in
QpointFrequencies