Temporarily restore tight_layout() to check CI results
This tight_layout() seems to be responsible for setting the
xticklabels of an invisible axis (used for label location) from empty
strings to numbers ranging 0 to 1. Those numbers are already in the
reference data so this was causing test failures.

We should consider removing it anyway, but it would be nice to get a
clean CI result first to isolate this from other changes.