Apply tight_layout() before comparing to script test data
A side-effect of tight_layout() is that it forces "unset" xticklabels
to take their values. (Presumably so that the sizes can be inspected
for collisions.) Now that this has been removed from plot_2d and
delegated to styling, the tick labels are not set at the point the
script test checks them. (Explicit labels such as high-symmetry points
have been, but not numbers, including the 0-1 range of the invisible
axis used to control the x-axis label position.)

The resulting test failures can be addressed by updating the reference
data to have these unset values; but perhaps it is more robust to
check the result of tight_layout(). The test does not see the final
rendered version of the plot, which would have similar labels.