Skip segfaulting tests on MacOS
Scipy > 1.1.0 from PyPI appears to unreliably segfault on
some MacOS machines (e.g. on Github MaOS runners but not on
PACE's Mac), and only segfaults on some CL tool tests, with
certain q-point paths/grids. There are quite a few workarounds
(using different --grid or --q-spacing arguments, or downgrading
Scipy) and it appears to be an issue with Scipy itself, so unless
this becomes a serious problem for users just skip these tests for
now. Possibly a similar issue is documented here
https://github.com/google/jax/issues/432