Move testing to Github Actions, closes #181 (#189)
* Add initial test run file

* Don't run on Jenkins CI

* Add Miniconda

* Don't abort jobs when one OS fails

* Increase tol on dispersion CL tests

* Test static analysis options

* Add release testing workflow

* Temporarily use push trigger in release test

* Run static analysis in specific directory

* Temporarily checkout current branch

Otherwise changes to release_tox won't be included

* Fix release_tox

- Add numpy as prerequisite
- Add absolute path to tox_requirements
- Remove index-url, not clear what it's needed for

* Parse test results

* Add conda release tests

* Remove pylint from CI requirements

* Whitelist rm command to avoid warning

* Use correct shell for Conda

* Set up release workflow dispatch

* Publish test results for all platforms

* Temporarily only test on mac

* Test structure factor without reciprocal ASR

ASR formalism is only valid at gamma, but the reciprocal
ASR gets applied at every q, so the problem is less
well-conditioned leading to higher, harder to test f.p.
errors on different systems. Instead largely test without
reciprocal ASR and increase tolerance in the cases where the
reciprocal ASR is used

* Increase tolerance on recip asr tests and add new no asr tests

* Correctly publish for multiple OS

* Remove additional ci requirements

* Skip segfaulting tests on MacOS

Scipy > 1.1.0 from PyPI appears to unreliably segfault on
some MacOS machines (e.g. on Github MaOS runners but not on
PACE's Mac), and only segfaults on some CL tool tests, with
certain q-point paths/grids. There are quite a few workarounds
(using different --grid or --q-spacing arguments, or downgrading
Scipy) and it appears to be an issue with Scipy itself, so unless
this becomes a serious problem for users just skip these tests for
now. Possibly a similar issue is documented here
https://github.com/google/jax/issues/432

* Run on all OSes and update release testing

* Remove static analysis action

This is run on Codacy rather than the Github Action runners

* Remove unused testing files

* Remove codecov action

* Avoid duplicate testing of pushes/PRs

* Run PR merge CI when 'run CI' label is used