Test styling: fontsize from style sheet
This uses the "small" font scaling for xticklabels with a specified
base font size.

This integration test should give a bit more confidence that the
styling system is capable of modifying text without raising an
error. The font names are going to be a bit more fragile (depending on
system fonts) so for now they are not tested.