Styling: add --style and --no-base-style to CLI plotters (#188)
* Styling: add --style and --no-base-style to CLI plotters

Matplotlib allows several stylesheets to be overlaid. To get the best
use of this, we should remove any explicit styling from plot functions
and replace them with style contexts. Then our settings can take a
higher priority than MPL defaults, but a lower priority than any user
style preferences. Styles can be selected by inbuilt names and from
user files.

This allows the size and formatting of plots to be customised in a
precise and consistent way for publication. It also simplifies
plotting functions by removing options that can be set by the context.

* Address the more reasonable Codacy failures

Are we really going to use CAPS for all module-level "constants"?

* Add style files to MANIFEST

* Add --fontsize, --linewidth options; tweak default font sizes

Have tweaked suptitle and tick labels to be bigger and smaller than
fontsize, respectively. I don't have a strong opinion on that, but Duc
asked for it indirectly!

* Add --font option

* Add styles directory to setup.py packages

Jenkins test runs are failing to import from euphonic.styles, maybe
it's not being included in the package correctly?

* Implement --figsize, --figsize-unit options

I never liked that Matplotlib uses inches as its natural length
unit. As we have access to an absurdly powerful unit conversion
engine, it would be rude not to use it.

Yes, this means you can specify your figure size in Bohr radii!

* Temporarily restore tight_layout() to check CI results

This tight_layout() seems to be responsible for setting the
xticklabels of an invisible axis (used for label location) from empty
strings to numbers ranging 0 to 1. Those numbers are already in the
reference data so this was causing test failures.

We should consider removing it anyway, but it would be nice to get a
clean CI result first to isolate this from other changes.

* Remove (invisible) x tick values from intensity map script test

* Revert "Remove (invisible) x tick values from intensity map script test"

This reverts commit 19586ec8223e8d121afb69d4d772a56018c829da.

* Apply tight_layout() before comparing to script test data

A side-effect of tight_layout() is that it forces "unset" xticklabels
to take their values. (Presumably so that the sizes can be inspected
for collisions.) Now that this has been removed from plot_2d and
delegated to styling, the tick labels are not set at the point the
script test checks them. (Explicit labels such as high-symmetry points
have been, but not numbers, including the 0-1 range of the invisible
axis used to control the x-axis label position.)

The resulting test failures can be addressed by updating the reference
data to have these unset values; but perhaps it is more robust to
check the result of tight_layout(). The test does not see the final
rendered version of the plot, which would have similar labels.

* New docs page: plot customisation

* Customisation docs tweaks

* Customisation docs correction.

The background colour hasn't changed, it's an illusion from the
missing axis outline! I checked the pixels with a colour picker.

* Styling: Tweaks in response to reviewer comments

Thanks @rebeccafair

* Spectrum docs: tweak subsection formatting

Sphinx is having trouble interpreting the cross-references to styling
file. Maybe the ---- for sections is incorrect and throwing it off?

* More RST wrangling

Looks like the highlight command was preventing the heading from being
identified. Fixed in all the script docs for consistency. (Somehow the
index was working anyway, but maybe cross-refs wouldn't?)

* Correct colour in styling docs alt-text

Good spot @mducle

* Add test for style arguments dispatch (internal function)

This ensures good coverage of the input space. We can check that the
appropriate results turn up in plotting outputs with a small
integration test.

* Simple integration test for styling

- Check that some CL options make it to the plot

* Test styling: fontsize from style sheet

This uses the "small" font scaling for xticklabels with a specified
base font size.

This integration test should give a bit more confidence that the
styling system is capable of modifying text without raising an
error. The font names are going to be a bit more fragile (depending on
system fonts) so for now they are not tested.

* Flake8 cleanup

* Plot styling: changelog, plot_1d docs

- Changelog update. (Maybe a bit too verbose?)
- Illustrate use of matplotlib.style.context in user docs for plot_1d

* Styling docs: tweak for more consistency with matplotlib docs

* Move plot styling notes to higher heading level.