Add multiple_extras test marker
This should be added to any test which needs multiple optional
dependencies e.g. to a test which has both pytest.mark.matplotlib
and pytest.mark.phonopy_reader. This allows
'pytest -m multiple_extras' to be used rather than the convoluted
'pytest -m "(phonopy_reader and matplotlib) or (phonopy_reader and brille) or (matplotlib and brille)"'