Increase test tolerances
Not sure why the results for this test have changed
(only on Windows) when no changes should've affected
it. Perhaps a change in the order that tests are run,
as unit conversions in Pint can be affected by caching.
Either way, the change is very small so unlikely to be
a problem.