Check type of x_tick_labels
In euphonic.cli.utils._get_tick_labels, the label indices
are found with np.where, which returns an array of np.int32
NOT int. np.int32 is not JSON serialisable so this was causing
an error on writing to JSON.

This could have been solved in _get_tick_labels, but as setting
the labels like this on a Spectrum object seems like a common
thing to do, I thought it was better to instead validate
x_tick_labels and convert to plain ints inside the class. This goes
against the rest of Euphonic, where things are only validated on class
creation, but given the use-case I feel it is necessary (and maybe
this is a sign we should be stricter on validation elsewhere in Euphonic)