Reduce tolerance on spglib symmetry
Partial solution to https://github.com/brille/brille/issues/63.
Spglib by default has a high symmetry tolerance, whereas Brille's
is much lower, meaning that spglib finds symmetry operations that
Brille disagrees with, so can't find the equivalent atoms. As
Brille's (as of 0.5.7) tolerance can't be tuned, just decrease the
tolerance for spglib.

I have tested for CaHgO2 and N2Al2 listed in the issue, and this
choice of symprec ensures things work in these cases, but without
reducing the number of symmetry operations available for other
structures (such as quartz, La2Zr2O7), which is what we want.

This issue may reoccur in the future if we test with more systems
and we may have to re-tune get_symmetry_dataset's arguments symprec
or angle_tolerance.