Detect repeated q-points and include them as candidates for labels
When importing QpointFrequencies/QpointPhononModes on a high-symmetry
path it is possible that a repeated non-Gamma q-point will not be
identified as a turning point. Conventionally these repeated points
are used to indicate points-of-interest in band structure
data (i.e. places to insert a vertical line and a label).

It seems harmless to check for these while looking for Gamma and
turning points and add them to the "qpt_has_label" list, and will
occasionally catch a missing point.