Castep .phonon import: allow repeat points to be averaged
The current .phonon import behaviour will over-weight LO-TO-split
Gamma points in .phonon files, because each of these is reported/read
with the full q-point "weight" of the nominal Gamma point. The user
can request an arbitrary number of splitting directions, so we can't
really make assumptions about how many there will be.

This implementation makes use of the 'q-pt=' index to identify
redundant points. This seems like the most reliable way of getting the
weights to add up to 1 as we only combine points that CASTEP really
does consider to be "the same entry".