Castep .phonon import: allow repeat points to be averaged (#221)
* Castep .phonon import: allow repeat points to be averaged

The current .phonon import behaviour will over-weight LO-TO-split
Gamma points in .phonon files, because each of these is reported/read
with the full q-point "weight" of the nominal Gamma point. The user
can request an arbitrary number of splitting directions, so we can't
really make assumptions about how many there will be.

This implementation makes use of the 'q-pt=' index to identify
redundant points. This seems like the most reliable way of getting the
weights to add up to 1 as we only combine points that CASTEP really
does consider to be "the same entry".

* Review comments: int cast id, use existing test data

- Casting ID to int changes the results; some duplicates were not
  being captured correctly in previous implementation.
- Use a customised test to avoid largely-duplicated test file and
  adjust weights manually.

* Rework repeated-qpt-averaging implementation

This is a bit more clean and compact; hopefully it is also more
efficient. We capture information about the array indices while
iterating through the lines and recording repeated points.

Counting values and averaging out weights no longer involves any
further indexing or lookup, and the array of qpt_ids is no longer
required.

* CASTEP reader: general flake8 cleanup

* Flake8 cleanup

Codacy is complaining, maybe this will help

* Castep .phonon q-point averaging: include a 3-way split

Make sure the averaging isn't somehow bugged to only divide by 2.

* Changelog update

* Give new kwarg name in changelog. Should make it easier to find!