Detect repeated q-points in band-structure data (#218)
* Detect repeated q-points and include them as candidates for labels

When importing QpointFrequencies/QpointPhononModes on a high-symmetry
path it is possible that a repeated non-Gamma q-point will not be
identified as a turning point. Conventionally these repeated points
are used to indicate points-of-interest in band structure
data (i.e. places to insert a vertical line and a label).

It seems harmless to check for these while looking for Gamma and
turning points and add them to the "qpt_has_label" list, and will
occasionally catch a missing point.

* Tidy up import statements

* Add unit test for "forced" high-symetry label by repeat q-pt

The formatting seems to be quite inconsistent between decimal and
fractional points, not sure this is being handled ideally.

Still, the test shows that we are succesfully detecting the repeat
point, which was the purpose of the exercise.

* Tweak labelling unit test to only use fractions

Float printing is inconsistent between versions of dependencies,
leading to spurious failure.

The inconsistent printing is a problem to address, but it's one for a
different PR.

* Update docstring for get_qpoint_labels to reflect new criteria

The private method was already updated, but this is the one that is
visible in the user docs!

* Changelog update