Fix convert_fc_phases test
convert_fc_phases only reindexes the force constants matrix, and
doesn't remove the dipole from the force constants matrix (whereas
the dipole is now removed when reading from phonopy files). So we
cannot use the same NaCl_force_constants.json file as we use in the
ForceConstants.from_phonopy tests. So instead use a new file
NaCl_dipole_force_constants.json, which is reindexed, but doesn't
have the dipole removed (this is actually the same as the original
NaCl_force_constants.json file before the dipole fix was introduced).