Instrument broadening from Spectrum1D: fix weighting
There seem to be two y-axis conventions at work; for "sparse" data
that has not been binned, the weights indicate a symmetry weighting or
possible neutron scattering intensity. For data in a spectrum, there
is also a weight for the bin width to compensate for bin-size effects.

The variable-width broadening machinery currently expects the
former (which makes sense for adaptive DOS broadening) and converts to
the latter (which is the expected form for spectrum
objects). Broadening _from_ a spectrum object therefore needs to
account for this.