Instrument broadening test: fix trouble with Numpy versions
Figure out which versions of the random generators both
a) work
b) don't raise deprecation warnings

for our range of supported numpy versions.

What's especially annoying is that RandomState isn't even the current
recommended way of doing this; according to Numpy docs the
best-practice is now `rng = np.random.default_rng(seed)`, with the
resulting Generator providing random() and integers() methods.

But that doesn't exist in older versions, so we'll probably want to
migrate this in future.