Instrument broadening: use numpy Polynomial and Quantities
Requiring a special Tuple is a _tiny_ bit awkward, but it's easier
than defining a Class for the same information!

The unit handling is extra-clunky to handle the minimum dependencies;
newer numpy/Pint combinations allow a lot more Numpy functions to be
applied directly on Quantity objects.

For now we have a new method on Spectrum1D, this might get refactored.