euphonic-dos: accept polynomial energy resolution
An interesting problem is that if we try to use the width-dependent
broadening function for fixed-width, the results don't quite agree.
It's not a big problem to avoid this, and probably more efficient.
But it's weird, shouldn't it just create two kernels, align one of
them to the exact width and then mix the results to 100% "lower"?
Is it the actual convolution implementation giving different
results (e.g. due to re-binning?)