Fixed-width Lorentzian broadening: make more exact
The existing scheme truncates the Lorentzian function to 3
sigma. However, truncating Lorentzians is a poor way of making them
more efficient as the tails die away very slowly. Other approaches are
available, but for now it is safest if this exact implementation
always evaluates the kernel over twice the bin range. This guarantees
no truncation.

For similar reasons the numerical normalisation might not be the best
idea. It may be more correct for a long tail to carry intensity out of
the plot range and reduce the overall visible magnitude. It doesn't
really cost more to calculate the theoretical normalisation constant,
so let's use that instead.