Lorentzian broadening: restore naive normalisation
This is not ideal; in the case of very large broadening it will tend
to overestimate peak intensities. However, evaluating the kernel with
"correct" scaling derived from bin widths can drastically overestimate
peak heights at narrow broadening width.

There are better ways of doing this; the width range can be
constrained, or histogram integrals can be worked out
properly. However, this is not generally done and can lead to results
that look strange compared to other implementations (including the
Scipy Gaussian broadening) that use simple normalisation. Hopefully we
can revisit this some time.