Use of Brille in Euphonic (#104)
* Add Brille to setup.py

* Add function to read symmetry from castep file

* Add initial Brille interpolator

* Add Debye Waller calculation to BrilleInterpolator

* Remove Debye-Waller calculation from Brille Interpolator

* Use more sensible defaults

* Add more sensible grid defaults

* Add BrilleInterpolator to __init__

* Add brille optional dependency to tests

* Small refactor, improvements and add simple tests

- Allow BrilleInterpolator to be created directly from a filled
  Brille grid, and move ease-of-use logic to .from_force_constants
  to allow more flexibility in the future and parity with the
  rest of Euphonic
- Add typing and docstrings
- Add some tests to check object creation (not testing numerical
  results yet)

* Add tests for optional import

* Small style improvements

* Update Brille dependency

* Correctly mark brille tests

Grid must be a fixture so it is not run unless marked
tests are run

* Don't install and test Brille on MacOS

The latest (0.5.7) version of Brille bundles a libomp.dylib file
which uses OpenMP 13. Versions 12 and above seem to be very buggy
on MacOS, see https://github.com/scikit-learn/scikit-learn issue
21182 and related issues.

This can be fixed by downloading a version 11 libomp.dylib (e.g. from
https://anaconda.org/conda-forge/llvm-openmp/11.1.0/download/osx-64/llvm-openmp-11.1.0-hda6cdc1_1.tar.bz2)
and replacing the one bundled in brille/.dylibs. However, this is
something that should probably be fixed in the Brille releases rather
than hacked into our CI so for now just work around it.

* Add multiple_extras test marker

This should be added to any test which needs multiple optional
dependencies e.g. to a test which has both pytest.mark.matplotlib
and pytest.mark.phonopy_reader. This allows
'pytest -m multiple_extras' to be used rather than the convoluted
'pytest -m "(phonopy_reader and matplotlib) or (phonopy_reader and brille) or (matplotlib and brille)"'

* Rename brille_interpolator.py to brille.py

For less verbose importing

* Add more tests

* Add missing multiple_extras test markers

* Add tests for passing of interpolation kwargs

* Add tests for BrilleInterpolator.calculate_qpoint_phonon_modes

* Add calculate_qpoint_frequencies and tests

* Add Brille options to euphonic-powder-map

* Use denser Brille grid for tests

* Add euphonic-powder-map tests with Brille

* Increase test tolerances

Not sure why the results for this test have changed
(only on Windows) when no changes should've affected
it. Perhaps a change in the order that tests are run,
as unit conversions in Pint can be affected by caching.
Either way, the change is very small so unlikely to be
a problem.

* Small updates from Codacy

* Small test changes:

- Fix qpts formatting in test
- More consistently name data files

* Reduce tolerance on spglib symmetry

Partial solution to https://github.com/brille/brille/issues/63.
Spglib by default has a high symmetry tolerance, whereas Brille's
is much lower, meaning that spglib finds symmetry operations that
Brille disagrees with, so can't find the equivalent atoms. As
Brille's (as of 0.5.7) tolerance can't be tuned, just decrease the
tolerance for spglib.

I have tested for CaHgO2 and N2Al2 listed in the issue, and this
choice of symprec ensures things work in these cases, but without
reducing the number of symmetry operations available for other
structures (such as quartz, La2Zr2O7), which is what we want.

This issue may reoccur in the future if we test with more systems
and we may have to re-tune get_symmetry_dataset's arguments symprec
or angle_tolerance.

* Increase number of qpts used in CaHgO2 test

* Use more consistent trellis rather than mesh in tests

* Add n-d recurrent sequences to euphonic.sampling and show-sampling

This sampling method produces points distributed across irrational
numbers, which should make a good reference set for interpolation
accuracy tests.

* Rough demo: Brille error checker

Add a CLI tool that plots interpolated vs calculated frequencies for a
mesh of quasirandom (irrationally-divided) q-points.

This plot could be replaced with some more useful metrics, but for now
is a starting-point to explore behaviour.

Script name will probably also change before we're done!

* Brille checking: nicer plot with multiple densities overlaid

* Add Brille grid_density and rename CL args for consistency

Use --brille-npts instead of --brille-n-qpts to match the powder
argument --npts and -npts-density

* Add structure factor comparison plot

* Add frequencies axis to structure factor plot

* Add summation over degenerate modes

* Add SQW and residual plots

* Remove structure factor summation

* Improve plotting logic

* Rename brille CL tool, and some small changes

- Remove unused imports
- Pass number of threads through to Brille

* Rename n_grid_points to grid_npts

* Update syntax for Brille 0.6

* Use grid_npts in brille-convergence

* Add ApproxConfig and snap_to_symmetry

* Add approx_config to enable CaHgO2 case

* Use kwargs in Lattice

* Use brille grids from file in tests

* Don't convert evecs to basis

* Clarify Brille vector elements input

* Use new Brille elements interface

* Remove references to basis eigenvectors

* Update test files (after new Brille version)

* Small Codacy updates

* Get line data from specific fig for testing

* Add initial brille-convergence script test

* Improve per-figure brille-convergence tests

* Add new convergence script test cases

* Ensure CL kwargs are passed through

* Test on CI with Mac again

* Fix tests on CI

- Increase tolerances
- Increase number of q-points in structure factor tests - otherwise
  results vary too much on different systems
- Don't use brille.ApproxConfig at top level (breaks for tests where
  Brille isn't installed)
- Close all plots in test_plot to avoid interference with
  brille-convergence tests
- Add alternative to get_data_3d (not available until Matplotlib 3.1)

* Small improvements to BrilleInterpolator

* Remove CaHgO2 test

Unstable structure factor values on different systems - hard to test

* Small Codacy updates

* Add user docs

* Add CL tool test case (multiple Brille npts)

* Remove ApproxConfig

snap_to_symmetry should be sufficient

* Minor reword of user docs

* Add Brille to readthedocs dependencies

Otherwise the BrilleInterpolator docstring isn't visible
as euphonic.brille can't be imported

* Update changelog

* Fix small doc typos

---------

Co-authored-by: Adam J. Jackson <a.j.jackson@physics.org>